requirements.txt
asgiref
distlib
Django
djangorestframework
djangorestframework-simplejwt
filelock
pip
platformdirs
PyJWT
python-dotenv
sqlparse
tzdata
virtualenv
